% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{euc_dist}
\alias{euc_dist}
\title{Eucledean distance
edge filter y; same as edge_filter_y but for y}
\usage{
euc_dist(x1, x2)
}
\description{
Eucledean distance
edge filter y; same as edge_filter_y but for y
}
