% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{get_contour_plot}
\alias{get_contour_plot}
\title{R package for microplastic image analysis
get_contour_plot}
\usage{
get_contour_plot(img, min_val = 0.1)
}
\arguments{
\item{img}{image source}

\item{min_val}{threshold}
}
\description{
R package for microplastic image analysis
get_contour_plot
}
\author{
Junah Bang \email{juna0033@g.skku.edu} and Youngdeok Hwang \email{yhwang@g.skku.edu}
}
