% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{get_curvature}
\alias{get_curvature}
\title{function to calculate the local curvature (EBImage vignettes)}
\usage{
get_curvature(img, h = 10)
}
\description{
function to calculate the local curvature (EBImage vignettes)
}
