% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{get_curvature_image}
\alias{get_curvature_image}
\title{function to assign the local curvature to the boundary for visulization}
\usage{
get_curvature_image(img, h = 10)
}
\description{
function to assign the local curvature to the boundary for visulization
}
