% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{get_edge}
\alias{get_edge}
\title{get the edge
edge filter y; same as edge_filter_y but for y}
\usage{
get_edge(img, min_val = 0.01, type)
}
\arguments{
\item{img}{image source}

\item{min_val}{threshold}
}
\description{
get the edge
edge filter y; same as edge_filter_y but for y
}
