% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{get_piece_dist}
\alias{get_piece_dist}
\title{calcuating the distance from the center tot he edge for each sample}
\usage{
get_piece_dist(edge_list, type)
}
\description{
calcuating the distance from the center tot he edge for each sample
}
