% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplastic.R
\name{piece_dist}
\alias{piece_dist}
\title{calculating the Eucledean distance from the center of the debris to the edge}
\usage{
piece_dist(x, y)
}
\description{
calculating the Eucledean distance from the center of the debris to the edge
}
